offer a basic initialization routine for test cases.
This commit generifies Invoke() and InvokeMap() functions to take
any of two test case function signatures:

- TestCaseFn (original)
- InitializedTestCaseFn (new)

When the latter is passed, we perform a basic initialization ceremony,
which is recurringly seen across test plans. Refer to godocs of that
type for more info.

This commit also moves the `runtime.Invoke*` methods into the `run`
package. This was necessary to resolve an import cycle (thanks Go!).

Finally, this commit makes the functions in the `network` package
no-op if we're running in a sidecar-less environment.